% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_apply.R
\name{df_apply}
\alias{df_apply}
\title{Apply functions to data frame}
\usage{
df_apply(.data, .f, .condition = is.numeric, .else = return, ...)
}
\arguments{
\item{.data}{A data frame}

\item{.f}{a function}

\item{.condition}{a condition function}

\item{.else}{a function to apply when .condition() evaluates to FALSE}

\item{...}{additional arguments that will be passed to .f()}
}
\value{
A data frame
}
\description{
Apply functions to data frame
}
\examples{
df_apply(iris, round, is.numeric, return, digit = 1)

}
